backend/pyproject.toml
[project]
name = "crew7-api"
version = "0.1.0"
dependencies = [
	"fastapi[all]",
	"uvicorn[standard]",
	"gunicorn>=21.2",
	"sqlalchemy>=2.0",
	"psycopg[binary]",
	"pydantic>=2.6",
	"pydantic-settings>=2.4",
	"redis>=5.0",
	"qdrant-client==1.9.2",
	"minio>=7.2",
	"python-multipart>=0.0.9",
	"httpx>=0.27",
	"passlib[bcrypt]>=1.7",
	"bcrypt<4",
	"pyjwt>=2.8",
	"email-validator>=2.0",
	"structlog>=23.1",
	"opentelemetry-api>=1.18",
	"opentelemetry-sdk>=1.18",
	"opentelemetry-exporter-otlp>=1.18",
	"prometheus-fastapi-instrumentator>=6.0",
	"prometheus-client>=0.16",
	"rq>=1.16",
	"docker==7.0.0",
	"gitpython==3.1.43",
	"stripe==10.12.0",
	"crewai==0.193.2",
	"langchain==0.2.15",
	"langchain-ollama==0.1.3",
	"langchain-community==0.2.10"
]


[project.optional-dependencies]
test = [
	"pytest>=8.3.2",
	"fakeredis>=2.23"
]
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
